4791 - move useHasOriginalDataPointData
