+++ src/client/pages/Section/Descriptions/hooks/useDescriptions.ts
@@ -1,7 +1,7 @@
 import { Description } from 'meta/assessment/description'
 
 import { useAssessmentCountry } from 'client/store/area/hooks/country'
-import { useHasOriginalDataPointData } from 'client/store/data/originalDataPoint/hooks/originalDataPoint'
+import { useHasOriginalDataPointData } from 'client/store/data/tableData/nodeValues/hooks/data'
 import { useIsPrintRoute } from 'client/hooks/useIsRoute'
 import { useSectionContext } from 'client/pages/Section/context'
