+++ src/client/store/data/originalDataPoint/hooks/originalDataPoint.ts
@@ -3,7 +3,6 @@ import { Dates } from 'utils/dates'
 
 import { ODPReservedYear, OriginalDataPoint } from 'meta/assessment/originalDataPoint'
 
-import { useOriginalDataPointData } from 'client/store/data/tableData/nodeValues/hooks/data'
 import { useAppSelector } from 'client/store/hooks'
 import { useCountryRouteParams } from 'client/hooks/useRouteParams'
 
@@ -43,4 +42,3 @@ export const useODPYears = (): { years: Array<number>; reservedYears: Array<numb
     reservedYears: reservedYears.map((reservedYear: ReservedYear) => reservedYear.year),
   }
 }
-export const useHasOriginalDataPointData = (): boolean => Object.keys(useOriginalDataPointData()).length > 0
